Implement a FragmentRegistry for registering named fragments.
This is just one implementation of a FragmentRegistry that could be
provided to InMemoryCache to fill in missing fragment declarations in
transformForLink.

Learning from past mistakes, we will avoid referring directly to this
particular implementation of FragmentRegistry within the InMemoryCache
implementation, since that forces the full FragmentRegistry
implementation to be bundled even when it is not used.